+++ src/pages/ProductList/FilteringCategory/FilteringCategory.scss
