+++ src/pages/ProductList/SearchResultList/SearchResultList.scss
