+++ src/pages/ProductList/SortingResult/SortingResult.scss
